EXAMPLES/UCP: format print send/recv result based on data type
When printing the send/recv data, it needs the data type to
decide how to print the data. Here, we change the first parameter
type in function print_result to be data_meta structure which
includes the data type.
Due to print_result is called in request_finalize, it also needs
to adjust its API.

Signed-off-by: Changcheng Liu <jerrliu@nvidia.com>