EXAMPLES/UCP: print recved segmented zero tail str in am mode
In AM mode, when sender use iov mode to send data, the am's cb
ucp_am_recv_callback_t handler will got all the received the data.
Due to the generated string are zero tailed, when the client
sends several strings in iov method, there will exist '\0' in the
middle of the recved buffer. So, it need to skip to the middle
'\0' and output all the received strings.

Signed-off-by: Changcheng Liu <jerrliu@nvidia.com>